# docs/swagger.yaml
definitions:
...
  entities.Category:
    properties:
      categoryId:
        type: integer
      iconUrl:
        type: string
      name:
        type: string
    type: object
...
  res.AllCategoriesRes:
    properties:
      categories:
        items:
          $ref: '#/definitions/entities.Category'
        type: array
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  res.CommonRes:
    properties:
      error:
        type: string
      message:
        type: string
      result: {}
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /addToCart/{id}:
    post:
      description: Add a dish to the user's cart
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added to cart
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Add to Cart
      tags:
      - User Cart
...
  /cart:
    get:
      description: Fetch the user's cart
      produces:
      - application/json
      responses:
        "200":
          description: Cart is empty
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: View Cart
      tags:
      - User Cart
  /cart/{id}/decrement:
    patch:
      description: Decrement the quantity of a dish in the user's cart
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully decremented cart item
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Decrement Cart Item
      tags:
      - User Cart
  /cart/{id}/deleteItem:
    delete:
      description: Delete a dish from the user's cart
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted cart item
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Delete Cart Item
      tags:
      - User Cart
...
  /cart/empty:
    delete:
      description: Empty the user's cart
      produces:
      - application/json
      responses:
        "200":
          description: Successfully emptied cart
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Empty Cart
      tags:
      - User Cart
  /categories:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/res.AllCategoriesRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      summary: Get all categories
      tags:
      - General
  /categories/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all categories
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/res.AllCategoriesRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      summary: Get all categories
      tags:
      - General
...
securityDefinitions:
  Bearer:
    description: 'Authentication using a JSON Web Token (JWT). The token should be
      included in the request header named "Authorization". The format of the header
      is: Authorization: Bearer <token>. Replace `<token>` with the actual JWT token.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"